refactor(grid): Move data mapper pipe
