+++ projects/igniteui-angular/src/lib/grids/grid/grid.module.ts
@@ -13,8 +13,7 @@ import {
     IgxGridPagingPipe,
     IgxGridGroupingPipe,
     IgxGridSortingPipe,
-    IgxGridFilteringPipe,
-    IgxGridDataMapperPipe
+    IgxGridFilteringPipe
 } from './grid.pipes';
 import { IgxGridGroupByRowComponent } from './groupby-row.component';
 import { IgxGridRowComponent } from './grid-row.component';
@@ -43,8 +42,7 @@ import { IgxGridExpandableCellComponent } from './expandable-cell.component';
     IgxGridFilteringPipe,
     IgxGridSummaryPipe,
     IgxGridDetailsPipe,
-    IgxGridExpandableCellComponent,
-    IgxGridDataMapperPipe
+    IgxGridExpandableCellComponent
   ],
   exports: [
     IgxGridComponent,
